reworked inventory and items now stack
I should commit more often. This commit has the updated stackable inventory with all its components working: droping only 1 item and making a copy of it on the map, staking in inventory, using 1 instance of items and removing it when it's no longer of quantity higher than 0. Also i restarted my IDE and had a bug be fixed :D